Move utility functions to utils module
