+++ sdgym/__main__.py
@@ -1,5 +1,6 @@
+"""SDGym CLI module."""
+
 import argparse
-import importlib
 import gc
 import logging
 import sys
@@ -15,23 +16,6 @@ import sdgym
 
 class SDGymError(Exception):
     pass
-
-
-def import_object(object_name):
-    """Import an object from its Fully Qualified Name."""
-
-    if isinstance(object_name, str):
-        parent_name, attribute = object_name.rsplit('.', 1)
-        try:
-            parent = importlib.import_module(parent_name)
-        except ImportError:
-            grand_parent_name, parent_name = parent_name.rsplit('.', 1)
-            grand_parent = importlib.import_module(grand_parent_name)
-            parent = getattr(grand_parent, parent_name)
-
-        return getattr(parent, attribute)
-
-    return object_name
 
 
 def _env_setup(logfile, verbose):
@@ -93,7 +77,7 @@ def _run(args):
                 synthesizers[synthesizer] = baselines[synthesizer]
             else:
                 try:
-                    synthesizers[synthesizer] = import_object(synthesizer)
+                    synthesizers[synthesizer] = sdgym.utils.import_object(synthesizer)
                 except Exception:
                     raise SDGymError(f'Unknown synthesizer {synthesizer}')
